requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-index-url --no-emit-trusted-host requirements.in
#
blessed==1.19.1
    # via -r requirements.in
grapheme==0.6.0
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
six==1.16.0
    # via blessed
wcwidth==0.2.5
    # via blessed
